Replace pendulum INLA demo with updated version
