+++ demos/pendulum_inla.py
@@ -1,34 +1,47 @@
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy import sparse
 from scipy.sparse.linalg import spsolve
 from scipy.integrate import odeint
-from sksparse.cholmod import cholesky
 from findiff import FinDiff, Coef, Identity
 
-from spdeinf import linear, nonlinear, util
+from spdeinf import util
+from spdeinf.nonlinear import SPDEDynamics, IterativeINLARegressor
+from spdeinf.distributions import LogNormal
 
-## Generate data from nonlinear damped pendulum eqn.
+np.random.seed(2)
+
+#####################################################
+# Generate data from nonlinear damped pendulum eqn. #
+#####################################################
 
 # Define parameters of damped pendulum
 b = 0.3
 c = 1.
 params_true = np.array([b, c])
 
-# Define parameters of the parameter priors
+# Define parameters of the model parameter priors
 tau_b = 1
 b_prior_mode = 0.2
 b_0 = np.log(b_prior_mode) + (tau_b ** (-2))
+
 tau_c = 1
 c_prior_mode = 2.
 c_0 = np.log(c_prior_mode) + (tau_c ** (-2))
-print(b_prior_mode, c_prior_mode)
-# param0 = np.array([b_prior_mode, c_prior_mode, 1, 1e-1])
-# param0 = np.array([b_prior_mode, c_prior_mode])
-param0 = np.array([b_prior_mode, 1])
-# param_bounds = [(0.1, 1), (0.1, 15), (1, 50), (1e-4, 1)]
-# param_bounds = [(0.1, 1), (0.1, 15)]
-param_bounds = [(0.1, 1), (1, 10)]
+
+# Process noise prior
+tau_k = 0.5
+k_prior_mode = 0.05
+k_0 = np.log(k_prior_mode) + (tau_k ** (-2))
+
+# Observation noise prior
+tau_s = 1
+s_prior_mode = 0.2
+s_0 = np.log(s_prior_mode) + (tau_s ** (-2))
+
+param0 = np.array([b_prior_mode, c_prior_mode, k_prior_mode, s_prior_mode])
+param_priors = [LogNormal(mu=b_0, sigma=1/tau_b), LogNormal(mu=c_0, sigma=1/tau_c),
+                LogNormal(mu=k_0, sigma=1/tau_k), LogNormal(mu=s_0, sigma=1/tau_s)]
+param_bounds = [(0.1, 1), (0.1, 5), (0, 1), (0, 1)]
 
 # Create temporal discretisation
 L_t = 25                      # Duration of simulation [s]
@@ -37,7 +50,7 @@ N_t = int(L_t / dt) + 1       # Points number of the temporal mesh
 T = np.linspace(0, L_t, N_t)  # Temporal array
 T = np.around(T, decimals=1)
 
-# Define the initial condition    
+# Define the initial condition
 u0 = [0.75 * np.pi, 0.]
 
 # Define corresponding system of ODEs
@@ -61,156 +74,109 @@ obs_idxs = np.array(list(obs_dict.keys()), dtype=int)
 obs_vals = np.array(list(obs_dict.values()))
 print("Number of observations:", obs_idxs.shape[0])
 
-################################
-# Diff. op. and mean generator #
-################################
+######################################
+# Define nonlinear pendulum dynamics #
+######################################
 
-def get_diff_op(u0, dt, params):
+class PendulumDynamics(SPDEDynamics):
     """
-    Constructs current linearised differential operator.
+    The parameters of the model are, in order:
+    0. b
+    1. c
+    2. process amplitude
+    3. observation noise
     """
-    # b, c, _, _ = params
-    # b, c = params
-    b, _ = params
-    partial_t = FinDiff(1, dt, 1)
-    partial_tt = FinDiff(1, dt, 2)
-    u0_cos = np.cos(u0)
-    diff_op = partial_tt + Coef(b) * partial_t + Coef(c * u0_cos) * Identity()
-    return diff_op
 
-def get_prior_mean(u0, diff_op_gen, params):
-    """
-    Calculates current prior mean.
-    """
-    # b, c, _, _ = params
-    # b, c = params
-    b, _ = params
-    u0_cos = np.cos(u0)
-    u0_sin = np.sin(u0)
-    diff_op = diff_op_gen(u0, params)
-    diff_op_mat = diff_op.matrix(u0.shape)
-    prior_mean = spsolve(diff_op_mat, (c * (u0 * u0_cos - u0_sin)).flatten())
-    return prior_mean.reshape(u0.shape)
+    def __init__(self, dt) -> None:
+        super().__init__()
+        self.dt = dt
 
-diff_op_gen = lambda u, params: get_diff_op(u, dt, params)
-prior_mean_gen = lambda u, params: get_prior_mean(u, diff_op_gen, params)
+    def get_diff_op(self, u0, params, **kwargs):
+        """
+        Construct current linearised differential operator.
+        """
+        b, c, _, _ = params
+        partial_t = FinDiff(1, self.dt, 1)
+        partial_tt = FinDiff(1, self.dt, 2)
+        u0_cos = np.cos(u0)
+        diff_op = partial_tt + Coef(b) * partial_t + Coef(c * u0_cos) * Identity()
+        return diff_op
 
-##################################
-# log-pdf of parameter posterior #
-##################################
+    def get_prior_precision(self, u0, params, **kwargs):
+        """
+        Calculate current prior precision.
+        """
+        diff_op_guess = self.get_diff_op(u0, params)
+        L = util.operator_to_matrix(diff_op_guess, u0.shape, interior_only=False)
+        prior_precision = (self.dt / params[2]**2) * (L.T @ L)
+        return prior_precision
 
-def _logpdf_marginal_posterior(b, c, Q_u, Q_uy, Q_obs, mu_u, mu_uy, obs_dict, shape, regularisation=1e-3):
-    # Unpack parameters
-    obs_idxs = np.array(list(obs_dict.keys()), dtype=int)
-    obs_idxs_flat = shape[1] * obs_idxs[:,0] + obs_idxs[:,1]
-    obs_vals = np.array(list(obs_dict.values()), dtype=float)
+    def get_prior_mean(self, u0, params, **kwargs):
+        """
+        Calculate current prior mean.
+        """
+        _, c, _, _ = params
+        u0_cos = np.cos(u0)
+        u0_sin = np.sin(u0)
+        diff_op = self.get_diff_op(u0, params)
+        diff_op_mat = diff_op.matrix(u0.shape)
+        prior_mean = spsolve(diff_op_mat, (c * (u0 * u0_cos - u0_sin)).flatten())
+        return prior_mean.reshape(u0.shape)
 
-    # Define reused consts.
-    log_2pi = np.log(2 * np.pi)
-    N = Q_obs.shape[0]
-    M = Q_u.shape[0]
+    def get_obs_noise(self, params, **kwargs):
+        """
+        Get observation noise (standard deviation).
+        """
+        return params[3]
 
-    # Perform matrix factorisation to compute log determinants
-    Q_u_chol = cholesky(Q_u + regularisation * sparse.identity(Q_u.shape[0]))
-    Q_u_logdet = Q_u_chol.logdet()
-    Q_uy_chol = cholesky(Q_uy)
-    Q_uy_logdet = Q_uy_chol.logdet()
-    Q_obs_chol = cholesky(Q_obs)
-    Q_obs_logdet = Q_obs_chol.logdet()
+dynamics = PendulumDynamics(dt)
 
-    # Compute b prior terms
-    if b > 0:
-        log_b = np.log(b)
-        log_p_b = np.log(tau_b) - log_b - 0.5 * ((tau_b * (log_b - b_0)) ** 2 + log_2pi) # log-normal prior
-    else:
-        log_p_b = float("-inf")
+#################################
+# Fit model with iterative INLA #
+#################################
 
-    # Compute c prior terms
-    if c > 0:
-        log_c = np.log(c)
-        log_p_c = np.log(tau_c) - log_c - 0.5 * ((tau_c * (log_c - c_0)) ** 2 + log_2pi) # log-normal prior
-    else:
-        log_p_c = float("-inf")
-
-    # Compute GMRF prior terms
-    diff_mu_uy_mu_u = mu_uy - mu_u
-    log_p_ut = 0.5 * (Q_u_logdet - diff_mu_uy_mu_u.T @ Q_u @ diff_mu_uy_mu_u - M * log_2pi)
-
-    # Compute obs model terms
-    diff_obs_mu_uy = obs_vals - mu_uy[obs_idxs_flat]
-    log_p_yut = 0.5 * (Q_obs_logdet - diff_obs_mu_uy.T @ Q_obs @ diff_obs_mu_uy - N * log_2pi)
-
-    # Compute full conditional terms
-    log_p_uyt = 0.5 * (Q_uy_logdet - M * log_2pi)
-
-    logpdf = log_p_b + log_p_c + log_p_ut + log_p_yut - log_p_uyt
-    return logpdf
-
-def logpdf_marginal_posterior(x, u0, obs_dict, diff_op_gen, prior_mean_gen, return_conditional_params=False, debug=False):
-    # Process args
-    obs_count = len(obs_dict.keys())
-
-    # Compute prior mean
-    prior_mean = prior_mean_gen(u0, x)
-
-    # Construct precision matrix corresponding to the linear differential operator
-    diff_op_guess = diff_op_gen(u0, x)
-    L = util.operator_to_matrix(diff_op_guess, u0.shape, interior_only=False)
-    # prior_precision = x[2] * (L.T @ L)
-    prior_precision = x[1] * (L.T @ L)
-    # prior_precision = L.T @ L
-
-    # Get "data term" of full conditional
-    res = linear._fit_gmrf(u, obs_dict, obs_std, prior_mean, prior_precision, calc_std=return_conditional_params,
-                         include_initial_cond=False, return_posterior_precision=True, regularisation=1e-5)
-
-    # Define prior and full conditional params
-    mu_u = prior_mean
-    Q_u = prior_precision
-    mu_uy = res['posterior_mean']
-    Q_uy = res['posterior_precision']
-    Q_obs = sparse.diags([obs_std ** (-2)], 0, shape=(obs_count, obs_count), format='csc')
-
-    if debug:
-        plt.figure()
-        plt.plot(prior_mean[0])
-        plt.show()
-
-    # Compute marginal posterior
-    logpdf = _logpdf_marginal_posterior(x[0], x[1], Q_u, Q_uy, Q_obs, mu_u.flatten(), mu_uy.flatten(), obs_dict, u0.shape)
-    # logpdf = _logpdf_marginal_posterior(x[0], c, Q_u, Q_uy, Q_obs, mu_u.flatten(), mu_uy.flatten(), obs_dict, u0.shape)
-    if return_conditional_params:
-        return logpdf, mu_uy, res['posterior_var']
-    return logpdf
-
-
-## Fit GP with linearised SODE prior from nonlinear pendulum eqn.
-
-# Perform iterative optimisation
 max_iter = 10
-model = nonlinear.NonlinearINLASPDERegressor(u, 1, dt, param0, diff_op_gen, prior_mean_gen, logpdf_marginal_posterior,
-                                             mixing_coef=0.5, param_bounds=param_bounds, sampling_evec_scales=[0.1, 0.05],
-                                             sampling_threshold=1)
-# model = nonlinear.NonlinearINLASPDERegressor(u, 1, dt, param0, diff_op_gen, prior_mean_gen, logpdf_marginal_posterior,
-#                                              mixing_coef=0.5, param_bounds=param_bounds, sampling_evec_scales=[0.1, 0.05, 0.1, 0.1],
-#                                              sampling_threshold=1)
-model.fit(obs_dict, obs_std, max_iter=max_iter, animated=True, calc_std=True, calc_mnll=True)
+parameterisation = 'natural' # 'moment' or 'natural'
+model = IterativeINLARegressor(u, dynamics, param0,
+                               mixing_coef=0.5,
+                               param_bounds=param_bounds,
+                               param_priors=param_priors,
+                               sampling_evec_scales=[0.1, 0.1, 0.1, 0.02],
+                               sampling_threshold=1)
+
+model.fit(obs_dict, obs_std, max_iter=max_iter, parameterisation=parameterisation, animated=False, calc_std=False, calc_mnll=True)
 iter_count = len(model.mse_hist)
 
-# Plot fit
-plt.figure(figsize=(3,3))
-plt.plot(T, u.squeeze(), "b", label="Ground truth")
-plt.plot(T, model.posterior_mean.squeeze(), "k", label="Posterior mean")
-plt.plot(T, model.posterior_mean.squeeze() + model.posterior_std.squeeze(), "--", color="grey", label="Posterior std. dev.")
-plt.plot(T, model.posterior_mean.squeeze() - model.posterior_std.squeeze(), "--", color="grey")
-plt.axvline(dt * obs_loc_1, color='grey', ls=':')
-plt.scatter(dt * obs_idxs[:,1], obs_vals, c="r", marker="x", label="Observations")
-plt.xlabel("$t$")
-plt.ylabel("$u$")
-plt.tight_layout()
-plt.savefig("figures/pendulum/pendulum_spde_inla_fit2.pdf")
-plt.show()
+############
+# Plot fit #
+############
 
-# Save animation
-print("Saving animation...")
-model.save_animation("figures/pendulum/pendulum_inla_iter_animation4.gif", fps=3)
+num_samples = 5000
+final_dist = model.marginal_dist_u_y
+samples = final_dist.sample(num_samples)
+samples = np.reshape(samples, (num_samples, -1))
+
+weights = np.ones(num_samples) / num_samples
+creds = [50, 60, 70, 80, 90, 95]
+
+list_of_intervals = []
+for i in range(samples.shape[1]):
+    intervals = util.cred_wt(samples[:,i], weights, creds)
+    list_of_intervals.append(intervals)
+
+plt.figure(figsize=(4,3))
+for cred in creds:
+    lower_lims = [interval[cred][0] for interval in list_of_intervals]
+    upper_lims = [interval[cred][1] for interval in list_of_intervals]
+    plt.fill_between(T, lower_lims, upper_lims, color='steelblue', alpha=1-cred/120, edgecolor='None')
+
+plt.plot(T, u[0], c='C1', linewidth=4)
+plt.plot(T, model.u0[0], c='k', linestyle='--', linewidth=2)
+plt.axvline(obs_loc_1 * dt, color='grey', ls=':', linewidth=2) 
+plt.scatter(obs_idxs[:,1] * dt, obs_vals, c='k', zorder=10)
+plt.xlabel("$t$", fontsize=16)
+plt.ylabel("$u$", fontsize=16)
+plt.xticks(fontsize=12)
+plt.yticks(fontsize=12)
+plt.tight_layout()
+plt.show()
